Добавил реализацию классов фигур Rectangle, Triangle, Ellipse в стиле TDD

+++ lw4/Shapes/Point.h
@@ -2,5 +2,5 @@
 
 struct Point
 {
-	unsigned x, y;
+	float x, y;
 };
